balance, consolidate: prevent infinite recursion
The mergerfs_control_file() functions of mergerfs.balance and
mergerfs.consolidate will recurse infinitely (up to Python's recursion
limit) when given a path which does not contain a file named '.mergerfs'
in any of its parent directories. mergerfs.dup does not have this issue.
Update the megerfs_control_file() functions in the affected commands to
match the version in mergerfs.dup.

Fixes #60